requirements.txt
PyQt5==5.15.10
openpyxl==3.1.2
pymupdf
pdfplumber
pillow
PyPDF2>=3.0.0
python-docx>=0.8.11
openai>=1.0.0
requests>=2.28.0